fix(eng-analytics): rename test_health views module to suite_health
